Linter errors fixed, New Game button
New game button 'partially' works.  It does seem to recognize the user
who is signed in, but it is not resetting the cells.  Most of the linter
errors have been resolved as well, especially on game.js.  The reason
my New Game button wasn't working is because it was listed as an ID
rather than a class on one of the files.